chore(extract): pin transformers, add setup-reranker target, warm Qwen3-Reranker in unit conftest
- Pin `transformers>=4.51` (Qwen3-Reranker model card minimum). The
  upper bound is intentionally absent — `mlx-embeddings` already
  forces transformers 5.x, and Qwen3-Reranker works on both. Locked
  to 5.9.0 in uv.lock.
- New `make setup-reranker` target pre-pulls the 0.6B weights into
  HF cache so the conftest session warmup serves hermetic tests
  without network. 4B opt-in via QWEN3_FULL=1, mirroring `setup-mlx`.
- New `_warm_qwen3_reranker` autouse session fixture in
  tests/unit/conftest.py, cross-platform (CPU device on every host).
- Switch model-load kwarg to `dtype=` (transformers 5 deprecates
  `torch_dtype`).

+++ src/auto_research/extract/rerank.py
@@ -145,7 +145,7 @@ def _ensure_qwen3_reranker_warmup(
     torch_dtype = torch.float16 if dtype == "fp16" else torch.float32
     try:
         tokenizer: Any = AutoTokenizer.from_pretrained(repo)
-        model: Any = AutoModelForCausalLM.from_pretrained(repo, torch_dtype=torch_dtype)
+        model: Any = AutoModelForCausalLM.from_pretrained(repo, dtype=torch_dtype)
     except Exception as exc:
         raise RuntimeError(
             f"Qwen3-Reranker {model_id!r} could not be loaded — likely a "
